+++ benchmarker/generate/course.go
@@ -3,17 +3,19 @@ package generate
 import (
 	"fmt"
 	"math/rand"
-	"strconv"
 	"strings"
-	"sync"
 	"sync/atomic"
-	"time"
 
 	"github.com/isucon/isucon11-final/benchmarker/model"
 )
 
 const (
 	majorCourseProb = 0.7
+)
+
+var (
+	liberalCode int32
+	majorCode   int32
 )
 
 var (
@@ -56,23 +58,6 @@ var (
 		"成績は出席と課題の提出状況により判断する。",
 	}
 )
-
-func randElt(arr []string) string {
-	return arr[rand.Intn(len(arr))]
-}
-
-var (
-	once          sync.Once
-	timeGenerator *timer
-
-	liberalCode int32
-	majorCode   int32
-)
-
-func init() {
-	rand.Seed(time.Now().UnixNano())
-	newTimer()
-}
 
 func courseDescription() string {
 	return randElt(courseDescription1) + randElt(courseDescription2)
@@ -150,106 +135,4 @@ func CourseParam(faculty *model.Faculty) *model.CourseParam {
 	} else {
 		return liberalCourseParam(faculty)
 	}
-}
-
-func Submission() *model.Submission {
-	title := "test title"
-	data := []byte{1, 2, 3}
-	return model.NewSubmission(title, data)
-}
-
-var (
-	classRooms = []string{
-		"H101", "S323", "S423", "S512", "S513", "W933", "M011",
-	}
-	classDescription1 = []string{
-		"課題は講義の内容について300字以下でまとめてください。",
-		"課題は講義の内容についてあなたが調べたことについて500字以上1000字以内でまとめてください。",
-		"課題は講義中に出題するクイズへの回答を提出してください。",
-	}
-)
-
-func classDescription(course *model.Course, part uint8) string {
-	var desc strings.Builder
-	switch rand.Intn(10) {
-	case 0:
-		desc.WriteString(fmt.Sprintf("%s 第%d回の講義です。", course.Name, part))
-	case 1:
-		desc.WriteString(fmt.Sprintf("%s 第%s回の講義です。", course.Name, convertKanjiNumbers(part)))
-	case 2:
-		desc.WriteString(fmt.Sprintf("%s の講義です。", course.Name))
-	case 3:
-		desc.WriteString(fmt.Sprintf("第%d回の講義です。", part))
-	case 4:
-		desc.WriteString(fmt.Sprintf("第%s回の講義です。", convertKanjiNumbers(part)))
-	}
-	switch rand.Intn(5) {
-	case 0:
-		desc.WriteString(fmt.Sprintf("今日のMTG-Room IDは %03d-%03d-%04d です。", rand.Intn(1000), rand.Intn(1000), rand.Intn(10000)))
-	case 1:
-		desc.WriteString(fmt.Sprintf("今日の講義室は %s です。", randElt(classRooms)))
-	}
-	desc.WriteString(randElt(classDescription1))
-	return desc.String()
-}
-
-var kanjiNumbers = []string{"〇", "一", "二", "三", "四", "五", "六", "七", "八", "九"}
-
-func convertKanjiNumbers(n uint8) string {
-	switch {
-	case n < 10:
-		return kanjiNumbers[n]
-	default:
-		return strconv.Itoa(int(n)) // fallback
-	}
-}
-
-func ClassParam(course *model.Course, part uint8) *model.ClassParam {
-	var title string
-	switch rand.Intn(5) {
-	case 0:
-		title = fmt.Sprintf("%s 第%d回講義", course.Name, part)
-	case 1:
-		title = fmt.Sprintf("%s 第%s回講義", course.Name, convertKanjiNumbers(part))
-	case 2:
-		title = fmt.Sprintf("第%d回講義", part)
-	case 3:
-		title = fmt.Sprintf("第%s回講義", convertKanjiNumbers(part))
-	case 4:
-		title = "新規講義"
-	}
-
-	desc := classDescription(course, part)
-	createdAt := GenTime()
-	return &model.ClassParam{
-		Title:     title,
-		Desc:      desc,
-		Part:      part,
-		CreatedAt: createdAt,
-	}
-}
-
-type timer struct {
-	base  int64 // unix time
-	count int64
-
-	mu sync.Mutex
-}
-
-func newTimer() {
-	once.Do(func() {
-		timeGenerator = &timer{
-			base:  time.Now().Unix(),
-			count: 0,
-			mu:    sync.Mutex{},
-		}
-	})
-}
-
-func GenTime() int64 {
-	timeGenerator.mu.Lock()
-	defer timeGenerator.mu.Unlock()
-
-	timeGenerator.count++
-	return timeGenerator.base + timeGenerator.count
 }
